cmt 59 [30/08]
Se consume webservice con seguridad (usr y pass) que se seteará en
config.properties. Se arreglan varios defectos (GIT) y se crean metodos
para el correcto manejo de los input numericos en presentacion.

+++ GestPedDsk/src/gpd/dominio/producto/Producto.java
@@ -16,7 +16,7 @@ public class Producto implements Serializable {
 	private String descripcion;
 	private Float stockMin;
 	private Unidad unidad;
-	private Integer cantUnidad;
+	private Float cantUnidad;
 	private AplicaIva aplIva;
 	private Double precio;
 	private Sinc sinc;
@@ -73,10 +73,10 @@ public class Producto implements Serializable {
 		this.unidad = unidad;
 	}
 	
-	public Integer getCantUnidad() {
+	public Float getCantUnidad() {
 		return cantUnidad;
 	}
-	public void setCantUnidad(Integer cantUnidad) {
+	public void setCantUnidad(Float cantUnidad) {
 		this.cantUnidad = cantUnidad;
 	}
